+++ rust/stackablectl/CHANGELOG.md
@@ -9,12 +9,23 @@ All notable changes to this project will be documented in this file.
 - Pass the product namespace as a templating variable `NAMESPACE` to manifests.
   This should unblock demos to run in all namespaces, as they can template the namespace e.g. for the FQDN of services ([#355]).
 
-### Fixes
+### Fixed
+
+- Prefix `ui-http` port endpoints with `http://`, as e.g. used by hbase-operator ([#368]).
+
+## [25.3.0] - 2025-03-27
+
+### Fixed
 
 - Use `rustls-native-certs` so that `stackablectl` can be used in environments with internal PKI ([#351]).
+- Use `heritage` label when looking up the `minio-console` stacklet ([#364]).
+- Improve tracing and log output ([#365]).
 
 [#351]: https://github.com/stackabletech/stackable-cockpit/pull/351
 [#355]: https://github.com/stackabletech/stackable-cockpit/pull/355
+[#364]: https://github.com/stackabletech/stackable-cockpit/pull/364
+[#365]: https://github.com/stackabletech/stackable-cockpit/pull/365
+[#368]: https://github.com/stackabletech/stackable-cockpit/pull/368
 
 ## [24.11.3] - 2025-01-27